Testing new users

+++ js/checklist.js
@@ -14,7 +14,38 @@ $(function () {
     const db = firebase.firestore();
 
     var logged = localStorage.getItem("logged");
+    var email = localStorage.getItem("email");
     // var logged = "true";
+    // var email = "rperezga@gmail.com"
+
+    var exist = false;
+
+    function newUser() {
+        db.collection("users").doc().set({
+            'email': localStorage.getItem("email"),
+            'Store one gallon of water per person per day.': false,
+            'Ready-to-eat canned meats, fruits, and vegetables.': false,
+            'High energy foods.': false,
+            'Vitamins.': false,
+            '(20) adhesive bandages, various sizes.': false,
+            'Aspirin or nonaspirin pain reliever.': false,
+            'Anti-diarrhea medication.': false,
+            'Mess kits, or paper cups, plates, and plastic utensils.': false,
+            'Battery-operated radio and extra batteries.': false,
+            'Flashlight and extra batteries.': false,
+            'Cash or traveler’s checks, change.': false,
+            'Non-electric can opener, utility knife.': false,
+            'Plastic storage containers.': false,
+            'Toilet paper, towelettes.': false,
+            'Soap, liquid detergent.': false,
+            'Feminine supplies.': false,
+            'Personal hygiene items.': false,
+            'Sturdy shoes or work boots.': false,
+            'Blankets or sleeping bags.': false,
+            'Hat and gloves.': false,
+            'Thermal underwear.': false
+        });
+    }
 
     var ready = true;
     if (ready == false) {
@@ -59,13 +90,35 @@ $(function () {
         provider = new firebase.auth.GoogleAuthProvider();
         firebase.auth().signInWithPopup(provider)
             .then(result => {
-                user = result.user;
-                localStorage.setItem("logged", true);
-                localStorage.setItem("user", user.displayName);
-                localStorage.setItem("email", user.email);
 
-                $("#login").attr("hidden", false);
-                $("#nologin").attr("hidden", true);
+                db.collection("users").where("email", "==", user.email).get().then(function (querySnapshot) {
+                    querySnapshot.forEach(function (doc) {
+                        if (doc) {
+                            exist = true;
+                        }
+                    });
+                })
+
+                setTimeout(function () {
+                    if (exist == true) {
+                        user = result.user;
+                        localStorage.setItem("logged", true);
+                        localStorage.setItem("user", user.displayName);
+                        localStorage.setItem("email", user.email);
+
+                        $("#login").attr("hidden", false);
+                        $("#nologin").attr("hidden", true);
+                    } else {
+                        newUser();
+                        user = result.user;
+                        localStorage.setItem("logged", true);
+                        localStorage.setItem("user", user.displayName);
+                        localStorage.setItem("email", user.email);
+
+                        $("#login").attr("hidden", false);
+                        $("#nologin").attr("hidden", true);
+                    }
+                }, 500)
             })
             .catch(console.log);
     });
@@ -83,11 +136,12 @@ $(function () {
     });
 
     db.collection("users").onSnapshot(doc => {
+        console.log("Longitud: " + doc.length)
         $("#dataPending").text("");
         $("#dataReady").text("");
 
         doc.forEach(function (doc) {
-            if (doc.data().email == "rperezga@gmail.com") {
+            if (doc.data().email == email) {
 
                 for (var key in doc.data()) {
                     var value = doc.data()[key]
@@ -137,58 +191,35 @@ $(function () {
                         $("#dataReady").append(rowElementReady);
                     }
                 }
-            }else{
-                // var newUser = db.collection("users").doc();
-                // newUser.set({ 
-                //     'email': localStorage.getItem("email"),
-                //     'Store one gallon of water per person per day.': false, 
-                //     'Ready-to-eat canned meats, fruits, and vegetables.': false, 
-                //     'High energy foods.': false, 
-                //     'Vitamins.': false ,
-                //     '(20) adhesive bandages, various sizes.': false, 
-                //     'Aspirin or nonaspirin pain reliever.': false, 
-                //     'Anti-diarrhea medication.': false, 
-                //     'Mess kits, or paper cups, plates, and plastic utensils.': false ,
-                //     'Battery-operated radio and extra batteries.': false ,
-                //     'Flashlight and extra batteries.': false ,
-                //     'Cash or traveler’s checks, change.': false ,
-                //     'Non-electric can opener, utility knife.': false ,
-                //     'Plastic storage containers.': false ,
-                //     'Toilet paper, towelettes.': false ,
-                //     'Soap, liquid detergent.': false ,
-                //     'Feminine supplies.': false ,
-                //     'Personal hygiene items.': false ,
-                //     'Sturdy shoes or work boots.': false ,
-                //     'Blankets or sleeping bags.': false ,
-                //     'Hat and gloves.': false ,
-                //     'Thermal underwear.': false
-                // });	
+            } else {
+
             }
         })
+
     });
 
-    function downloadToCsv (json) {
+    function downloadToCsv(json) {
 
         var csvContent = "data:text/csv;charset=utf-8,";
- 
-        json.forEach(function(rowArray){
+
+        json.forEach(function (rowArray) {
             var row = rowArray.join(",");
             csvContent += row + "\r\n";
         });
- 
+
         var encodedUri = encodeURI(csvContent);
         window.open(encodedUri);
     }
- 
+
     $("#downloadBtn").on("click", function (event) {
         event.preventDefault();
- 
+
         var json = [["can", "city1", "some other info"], ["name2", "city2", "more info"]];
         downloadToCsv(json);
     });
 
 
-    
+
